refactor(view): extract Label + visualizers from widgets.cpp (R2-6)
widgets.cpp was 2,081 lines with 20+ touches in the last 60 days,
producing constant cross-widget merge conflicts. Two clean
extractions land here:

* `core/view/src/widgets/label.cpp` (~663 lines) — the Label class
  was by itself almost 700 lines (the biggest single widget by far,
  because the CSS-style typography cascade lives there).
* `core/view/src/widgets/visualizers.cpp` (~788 lines) — all the
  visualization-ish widgets: ImageView, Meter, XYPad, WaveformView,
  SpectrumView, Panel, SpectrogramView, MultiMeter, CorrelationMeter.

widgets.cpp itself drops from 2,081 → 835 lines (-60%), keeping the
schema renderer + animation helpers + Knob/Fader/RangeSlider/Toggle/
Checkbox/ToggleButton/Icon. Per Codex's R2-6 risk callout, no header
changes — every public declaration stays in pulp/view/widgets.hpp.

The "<300 lines per widget" target from the original R2-6 sketch is
explicitly dropped (Codex's correct objection — Label paint alone is
substantial). The kept validation is the conflict-surface benefit:
a Label-only PR now touches widgets/label.cpp and nothing else; a
Meter-only PR touches widgets/visualizers.cpp.

Validation:
* `pulp-view` builds clean (Release config).
* `pulp-test-widgets` passes 84/84 (373 assertions).

This is item 2 of Phase 2 (Round 2 Batch 3).
