+++ core/view/src/widgets.cpp
@@ -245,641 +245,8 @@ void Toggle::advance_animations(float dt) {
     hover_opacity_.advance(dt);
 }
 
-// ── Label ────────────────────────────────────────────────────────────────────
-
-float Label::intrinsic_height() const {
-    // issue-969: cascade font_size before computing height so descendants
-    // of a parent that called setInheritableFontSize report a height that
-    // actually matches what paint() will draw.
-    float effective_font_size = font_size_;
-    if (!has_own_font_size_) {
-        if (auto inh = inheritable_font_size(); inh.has_value())
-            effective_font_size = inh.value();
-    }
-    // pulp-internal #76 — small font sizes need a larger line-height
-    // multiplier than the standard 1.4. At fs=10 (Spectr's
-    // `<span fontSize=10>SNAPSHOT</span>` in the bottom toolbar), Inter's
-    // typographic ascent + descent is ~13px, leaving only ~1px of slack
-    // in a 14px box (10 * 1.4); the GPU clip-rect on the View bounds
-    // then shaved the descender-side slack off in practice and the label
-    // visibly clipped from the bottom. Bumping to 1.6 for sizes below
-    // ~12pt buys 16px of box room for fs=10, fully containing the glyph
-    // extent. Larger sizes keep 1.4 — they have plenty of absolute slack.
-    const float lh_mult = effective_font_size < 12.0f ? 1.6f : 1.4f;
-    const float lh = line_height_ > 0 ? line_height_ : effective_font_size * lh_mult;
-
-    // pulp-internal #74 — when the Label is multi_line, the reserved
-    // height must reflect the number of lines paint() will emit, not a
-    // hard-coded one-line metric. Otherwise Yoga reserves only `lh` of
-    // vertical room and the parent (overflow:hidden on the toolbar / row
-    // gap on Settings-modal section subtitles) clips every line after
-    // the first. Spectr's Settings-modal description paragraphs are the
-    // motivating case.
-    //
-    // The \n count is the lower bound here. Soft-wrap (no \n but text
-    // exceeds the available width) needs the width Yoga passes to the
-    // measure callback — see `measured_height(available_width)` below.
-    //
-    // Single-line labels (multi_line_ == false) keep the legacy one-line
-    // return so single-line widths/heights match exactly what paint()
-    // computes for `text_h = effective_font_size` (the contract every
-    // existing test depends on).
-    if (multi_line_ && !text_.empty()) {
-        int line_count = 1;
-        for (char c : text_) {
-            if (c == '\n') ++line_count;
-        }
-        // pulp #1969 Codex P2 — don't reserve a phantom line for a trailing
-        // newline. Label::paint()'s `\n`-split loop emits one line per
-        // *non-trailing* `\n` plus the final segment, so `"Title\n"` paints
-        // exactly one visible line. If we keep the naive `\n`-count + 1
-        // here, Yoga reserves extra vertical whitespace that paint never
-        // fills, breaking CSS-style vertical-align centering and shifting
-        // siblings down. Matches the line-box counting CSS uses for
-        // `white-space: pre`.
-        if (text_.back() == '\n') --line_count;
-        // Honor line-clamp if explicitly set — paint() will only emit
-        // `line_clamp_` lines, so reserving more height is wasteful and
-        // confuses CSS vertical-align centering.
-        if (line_clamp_ > 0 && line_clamp_ < line_count)
-            line_count = line_clamp_;
-        return lh * static_cast<float>(line_count);
-    }
-    return lh;
-}
-
-float Label::measured_height(float available_width) const {
-    // pulp-internal #74 — width-aware height for multi_line Labels with
-    // soft-wrap. The Yoga measure callback receives the available width
-    // during layout, which is the only place we can run the shaper to
-    // figure out how many lines a soft-wrap block will actually produce.
-    // Without this hook, Yoga reserves exactly one line `lh` and any
-    // wrapped line past the first paints into sibling territory and is
-    // visually clipped (Spectr Settings modal section subtitles, any
-    // dropdown label, any flex-row chrome that uses bounded-width
-    // multi-line text).
-    //
-    // Contract:
-    //   • single-line labels                 → intrinsic_height() (legacy).
-    //   • multi-line + zero/unbounded width  → intrinsic_height() (\n only).
-    //   • multi-line + finite width          → shaper line count * lh.
-    //
-    // The shaper itself is the same one paint() uses, so the count we
-    // return here matches what paint() draws — no off-by-one, no
-    // double-shape penalty (TextShaper::prepare() caches per
-    // (text, family, size); paint will hit the same cache entry).
-    if (!multi_line_ || text_.empty() || available_width <= 0.0f)
-        return intrinsic_height();
-
-    float effective_font_size = font_size_;
-    if (!has_own_font_size_) {
-        if (auto inh = inheritable_font_size(); inh.has_value())
-            effective_font_size = inh.value();
-    }
-    // Match intrinsic_height's small-font multiplier (pulp-internal #76)
-    // so the measured line height is consistent with what paint() draws.
-    const float lh_mult = effective_font_size < 12.0f ? 1.6f : 1.4f;
-    const float lh = line_height_ > 0 ? line_height_ : effective_font_size * lh_mult;
-
-    // Mirror paint()'s text-transform — the line count for an
-    // ALL-CAPS-via-text-transform string can differ from the source
-    // string's because uppercase advances are typically wider.
-    std::string display_text = text_;
-    if (text_transform_ == TextTransform::uppercase) {
-        for (auto& ch : display_text)
-            ch = static_cast<char>(std::toupper(static_cast<unsigned char>(ch)));
-    } else if (text_transform_ == TextTransform::lowercase) {
-        for (auto& ch : display_text)
-            ch = static_cast<char>(std::tolower(static_cast<unsigned char>(ch)));
-    } else if (text_transform_ == TextTransform::capitalize) {
-        bool cap_next = true;
-        for (auto& ch : display_text) {
-            if (cap_next && std::isalpha(static_cast<unsigned char>(ch))) {
-                ch = static_cast<char>(std::toupper(static_cast<unsigned char>(ch)));
-                cap_next = false;
-            }
-            if (ch == ' ') cap_next = true;
-        }
-    }
-
-    const std::string& family = font_family_.empty() ? std::string("Inter") : font_family_;
-    auto& shaper = canvas::global_text_shaper();
-    auto prepared = shaper.prepare(display_text, family, effective_font_size);
-
-    // Use the same break_mode pulp paint uses (CSS word-break / overflow-
-    // wrap; Label paint reads `View::word_break()` at draw time, the
-    // measure path mirrors that decision).
-    const std::string wb = word_break();
-    canvas::BreakMode break_mode = canvas::BreakMode::normal;
-    if      (wb == "break-word") break_mode = canvas::BreakMode::break_word;
-    else if (wb == "anywhere")   break_mode = canvas::BreakMode::anywhere;
-    auto layout = shaper.layout(prepared, available_width, lh, /*max_lines=*/0, break_mode);
-
-    int line_count = std::max(1, layout.line_count);
-    if (line_clamp_ > 0 && line_clamp_ < line_count)
-        line_count = line_clamp_;
-    return std::ceil(lh * static_cast<float>(line_count));
-}
-
-float Label::intrinsic_width() const {
-    // issue-928: report the natural shaped-text width so Yoga reserves
-    // enough horizontal space for the full label content. Without this,
-    // long labels in flex-row containers (e.g. Spectr's "ZOOMABLE FILTER
-    // BANK" header) inherit a small parent width and clip mid-word.
-    //
-    // For multi-line labels we deliberately return 0 so the parent
-    // container's available width drives line wrapping instead of the
-    // single-line text width.
-    if (text_.empty() || multi_line_) return 0;
-
-    // issue-969: intrinsic measurement must match what paint() will
-    // actually draw, so honor the same own→inherited cascade for
-    // font_size and letter_spacing.
-    float effective_font_size = font_size_;
-    if (!has_own_font_size_) {
-        if (auto inh = inheritable_font_size(); inh.has_value())
-            effective_font_size = inh.value();
-    }
-    float effective_letter_spacing = letter_spacing_;
-    if (!has_own_letter_spacing_) {
-        if (auto inh = inheritable_letter_spacing(); inh.has_value())
-            effective_letter_spacing = inh.value();
-    }
-
-    // pulp #943 P2 / #945: when paint() rotates the text 90° the
-    // horizontal footprint is just the line height, not the shaped
-    // string advance. Reporting the advance here makes Yoga reserve
-    // far too much width for vertical labels and starves siblings.
-    bool vertical = (text_direction_ == canvas::TextDirection::top_to_bottom ||
-                     text_direction_ == canvas::TextDirection::bottom_to_top);
-    if (vertical) {
-        // pulp-internal #76 — same small-font-size bump as intrinsic_height.
-        const float lh_mult = effective_font_size < 12.0f ? 1.6f : 1.4f;
-        return std::ceil(line_height_ > 0 ? line_height_ : effective_font_size * lh_mult);
-    }
-
-    // Mirror paint()'s text-transform — measurement must match what's
-    // drawn or a row of uppercase chars will wrap unexpectedly.
-    std::string display_text = text_;
-    if (text_transform_ == TextTransform::uppercase) {
-        for (auto& ch : display_text)
-            ch = static_cast<char>(std::toupper(static_cast<unsigned char>(ch)));
-    } else if (text_transform_ == TextTransform::lowercase) {
-        for (auto& ch : display_text)
-            ch = static_cast<char>(std::tolower(static_cast<unsigned char>(ch)));
-    } else if (text_transform_ == TextTransform::capitalize) {
-        bool cap_next = true;
-        for (auto& ch : display_text) {
-            if (cap_next && std::isalpha(static_cast<unsigned char>(ch))) {
-                ch = static_cast<char>(std::toupper(static_cast<unsigned char>(ch)));
-                cap_next = false;
-            }
-            if (ch == ' ') cap_next = true;
-        }
-    }
-
-    // Shape with the same font + size the painter will use. TextShaper
-    // uses the global Skia/HarfBuzz path when available and falls back
-    // to a character-width estimator otherwise — same fallback that
-    // Canvas::measure_text() uses on the recording / non-Skia backends.
-    auto& shaper = canvas::global_text_shaper();
-    auto prepared = shaper.prepare(display_text, "Inter", effective_font_size);
-    float width = prepared.total_width();
-
-    // Letter-spacing adds extra advance per glyph break that isn't
-    // captured by HarfBuzz shaping. Count UTF-8 *code points*, not
-    // bytes — using `size()` over-applies spacing on multibyte input
-    // (CJK, accented Latin, emoji) and inflates intrinsic width
-    // (pulp #943 P2 #935 finding 2).
-    if (effective_letter_spacing != 0 && !display_text.empty()) {
-        std::size_t glyph_count = 0;
-        for (unsigned char c : display_text) {
-            // Count any byte that is not a UTF-8 continuation byte
-            // (0b10xxxxxx) — that's one glyph per code point.
-            if ((c & 0xC0) != 0x80) ++glyph_count;
-        }
-        if (glyph_count > 1) {
-            width += effective_letter_spacing * static_cast<float>(glyph_count - 1);
-        }
-    }
-
-    // Sub-pixel-safe ceil so layout never clips on rounding.
-    return std::ceil(width);
-}
-
-void Label::paint(canvas::Canvas& canvas) {
-    if (text_.empty()) return;
-
-    // issue-969: CSS-style typography cascade. For each property:
-    //   1. Use the Label's own value if explicitly set.
-    //   2. Otherwise walk up the parent chain via View::inheritable_*().
-    //   3. Otherwise fall back to the existing theme/default behavior.
-    canvas::Color text_color;
-    if (has_own_text_color_) {
-        text_color = text_color_;
-    } else if (auto inherited = inheritable_text_color(); inherited.has_value()) {
-        text_color = inherited.value();
-    } else {
-        text_color = resolve_color("text.primary", canvas::Color::rgba8(200, 200, 200));
-    }
-    canvas.set_fill_color({text_color.r, text_color.g, text_color.b, text_color.a});
-
-    float effective_font_size = font_size_;
-    if (!has_own_font_size_) {
-        if (auto inh = inheritable_font_size(); inh.has_value())
-            effective_font_size = inh.value();
-    }
-    int effective_font_weight = font_weight_;
-    if (!has_own_font_weight_) {
-        if (auto inh = inheritable_font_weight(); inh.has_value())
-            effective_font_weight = inh.value();
-    }
-    float effective_letter_spacing = letter_spacing_;
-    if (!has_own_letter_spacing_) {
-        if (auto inh = inheritable_letter_spacing(); inh.has_value())
-            effective_letter_spacing = inh.value();
-    }
-
-    // pulp #927 — propagate setFontFamily / setFontWeight / setLetterSpacing
-    // through to the canvas backend so JS calls actually change rasterised
-    // glyphs. Empty font_family_ falls back to the default theme face.
-    const std::string& family = font_family_.empty() ? std::string("Inter") : font_family_;
-    canvas.set_font_full(family, effective_font_size, effective_font_weight,
-                          font_style_, effective_letter_spacing);
-
-    // pulp #1737 — translate CSS font-variant CSV → SkShaper Feature
-    // tags. Wires the storage-only View::font_variant_ slot into the
-    // active paint. Empty CSV → clear features so previous view's
-    // settings don't bleed across paint calls. Each CSS keyword maps
-    // to its OpenType feature tag (per CSS Fonts Module 4 §7.3):
-    //   tabular-nums       → tnum
-    //   small-caps         → smcp
-    //   oldstyle-nums      → onum
-    //   lining-nums        → lnum
-    //   proportional-nums  → pnum
-    // Unknown values are silently ignored (forward-compat).
-    const std::string& fv = font_variant();
-    if (fv.empty()) {
-        canvas.clear_font_features();
-    } else {
-        std::vector<canvas::Canvas::FontFeature> features;
-        size_t i = 0;
-        while (i < fv.size()) {
-            while (i < fv.size() && (std::isspace(static_cast<unsigned char>(fv[i])) || fv[i] == ',')) ++i;
-            if (i >= fv.size()) break;
-            size_t end = i;
-            while (end < fv.size() && fv[end] != ',') ++end;
-            std::string token(fv, i, end - i);
-            while (!token.empty() && std::isspace(static_cast<unsigned char>(token.back()))) token.pop_back();
-            if      (token == "tabular-nums")      features.push_back({canvas::Canvas::make_font_feature_tag("tnum"), 1});
-            else if (token == "small-caps")        features.push_back({canvas::Canvas::make_font_feature_tag("smcp"), 1});
-            else if (token == "oldstyle-nums")     features.push_back({canvas::Canvas::make_font_feature_tag("onum"), 1});
-            else if (token == "lining-nums")       features.push_back({canvas::Canvas::make_font_feature_tag("lnum"), 1});
-            else if (token == "proportional-nums") features.push_back({canvas::Canvas::make_font_feature_tag("pnum"), 1});
-            // Unknown token → silently ignored.
-            i = end + 1;
-        }
-        if (!features.empty()) canvas.set_font_features(std::move(features));
-        else                    canvas.clear_font_features();
-    }
-
-    // Apply text-transform
-    std::string display_text = text_;
-    if (text_transform_ == TextTransform::uppercase) {
-        for (auto& ch : display_text) ch = static_cast<char>(std::toupper(static_cast<unsigned char>(ch)));
-    } else if (text_transform_ == TextTransform::lowercase) {
-        for (auto& ch : display_text) ch = static_cast<char>(std::tolower(static_cast<unsigned char>(ch)));
-    } else if (text_transform_ == TextTransform::capitalize) {
-        bool cap_next = true;
-        for (auto& ch : display_text) {
-            if (cap_next && std::isalpha(static_cast<unsigned char>(ch))) {
-                ch = static_cast<char>(std::toupper(static_cast<unsigned char>(ch)));
-                cap_next = false;
-            }
-            if (ch == ' ') cap_next = true;
-        }
-    }
-
-    // Vertical text direction — rotate canvas for top-to-bottom / bottom-to-top
-    bool vertical = (text_direction_ == canvas::TextDirection::top_to_bottom ||
-                     text_direction_ == canvas::TextDirection::bottom_to_top);
-    if (vertical) {
-        canvas.save();
-        if (text_direction_ == canvas::TextDirection::top_to_bottom) {
-            canvas.translate(bounds().width * 0.5f + effective_font_size * 0.35f, 0);
-            canvas.rotate(3.14159265f / 2.0f);
-        } else {
-            canvas.translate(bounds().width * 0.5f - effective_font_size * 0.35f, bounds().height);
-            canvas.rotate(-3.14159265f / 2.0f);
-        }
-    }
-
-    // Vertical alignment
-    // pulp-internal #76 — match intrinsic_height's small-font-size bump so
-    // multi-line / shaper-wrapped layout uses the same line-box height
-    // that Yoga reserved. Without matching here, an fs=10 multi-line
-    // label would size to 16px boxes but paint at 14px line height and
-    // siblings would not align.
-    const float lh_mult = effective_font_size < 12.0f ? 1.6f : 1.4f;
-    float lh = line_height_ > 0 ? line_height_ : effective_font_size * lh_mult;
-
-    // pulp #1737 PR-2 / #1924 — CSS `white-space` / `overflow-wrap` /
-    // `word-break` soft-wrap path. Route any multi-line, bounded-width
-    // Label through TextShaper so CSS default `white-space: normal`
-    // soft-wraps at word boundaries (whitespace), and `break-word` /
-    // `anywhere` additionally split inside over-wide words.
-    //
-    // Pre-#1924 this gate also required `break_mode != normal`, which
-    // meant the default mode skipped the shaper and never soft-wrapped —
-    // dropdown labels (and any other multi_line Label without an
-    // explicit word-break) overflowed their container instead of
-    // wrapping. TextShaper's `BreakMode::normal` already preserves the
-    // legacy "whole-word overflow for a single unbroken word" behavior
-    // (see test_text_shaper.cpp [issue-1737]), so removing the gate
-    // brings Label in line with the CSS spec without regressing the
-    // single-long-word case.
-    //
-    // The shaped layout is computed ONCE here and reused both for the
-    // vertical-align metrics (source_lines / text_h) and the rendering
-    // loop further down. This avoids a double-shape and keeps line-clamp /
-    // ellipsis / vertical-align integration coherent across modes.
-    const std::string wb = word_break();
-    canvas::BreakMode break_mode = canvas::BreakMode::normal;
-    if      (wb == "break-word") break_mode = canvas::BreakMode::break_word;
-    else if (wb == "anywhere")   break_mode = canvas::BreakMode::anywhere;
-    const bool use_shaper_wrap =
-        multi_line_ &&
-        bounds().width > 0.0f;
-
-    canvas::ShapedLayout shaped_layout;
-    if (use_shaper_wrap) {
-        auto& shaper = canvas::global_text_shaper();
-        auto prepared = shaper.prepare(display_text, family, effective_font_size);
-        shaped_layout = shaper.layout_with_lines(
-            prepared, bounds().width, lh, /*max_lines=*/0, break_mode);
-    }
-
-    // pulp #1552 — when line-clamp drops source lines, the painted block
-    // height must reflect the *visible* line count, not the full newline
-    // count. Otherwise vertical-align: center / bottom positions the
-    // block as if the hidden lines were still rendered, leaving the
-    // visible lines offset upward (Codex P2 on PR #1573).
-    //
-    // pulp #1737 PR-2 — `source_lines` for the soft-wrap path is the
-    // shaped layout's line count (which already accounts for inside-word
-    // breaks). For the legacy path it stays count('\n') + 1.
-    //
-    // pulp #1969 Codex P2 — drop a trailing newline before counting in the
-    // legacy path. The split-and-emit loop below stops once `pos ==
-    // display_text.size()`, so a trailing `\n` doesn't actually paint an
-    // extra line — but feeding the inflated count into `text_h` would
-    // mis-position vertical-align: center/bottom. The shaper path doesn't
-    // need a fix here because shaped_layout.line_count is the count the
-    // shaper actually produced.
-    int source_lines = multi_line_
-        ? (use_shaper_wrap
-               ? std::max(1, shaped_layout.line_count)
-               : static_cast<int>(std::count(display_text.begin(),
-                                             display_text.end(), '\n')) + 1
-                 - (!display_text.empty() && display_text.back() == '\n' ? 1 : 0))
-        : 1;
-    int visible_lines = source_lines;
-    if (multi_line_ && line_clamp_ > 0 && line_clamp_ < source_lines)
-        visible_lines = line_clamp_;
-    float text_h = multi_line_ ? lh * static_cast<float>(visible_lines) : effective_font_size;
-    float baseline_y;
-    switch (vertical_align_) {
-        case canvas::TextVerticalAlign::top:
-            baseline_y = effective_font_size * 0.85f;
-            break;
-        case canvas::TextVerticalAlign::bottom:
-            baseline_y = bounds().height - text_h + effective_font_size * 0.85f;
-            break;
-        case canvas::TextVerticalAlign::baseline:
-            baseline_y = bounds().height * 0.75f;
-            break;
-        case canvas::TextVerticalAlign::center:
-        default:
-            // Centre the visible block within bounds, then offset to the
-            // first line's baseline. For single-line this collapses to
-            // bounds.h/2 + 0.35*font_size (the historic formula) because
-            // text_h == effective_font_size and 0.85 - 0.5 == 0.35.
-            baseline_y = (bounds().height - text_h) * 0.5f + effective_font_size * 0.85f;
-            break;
-    }
-
-    // issue-969: text-align cascade. Own value wins, otherwise inherited.
-    LabelAlign effective_text_align = text_align_;
-    if (!has_own_text_align_) {
-        if (auto inh = inheritable_text_align(); inh.has_value()) {
-            int v = inh.value();
-            if (v == 1) effective_text_align = LabelAlign::center;
-            else if (v == 2) effective_text_align = LabelAlign::right;
-            else if (v == 3) effective_text_align = LabelAlign::auto_;
-            else if (v == 4) effective_text_align = LabelAlign::justify;
-            else if (v == 5) effective_text_align = LabelAlign::match_parent;
-            else effective_text_align = LabelAlign::left;
-        }
-    }
-
-    // pulp #1434 — resolve `match-parent` at paint time. CSS spec: the
-    // computed value matches the parent's resolved `text-align`. We walk
-    // the View parent chain via `inheritable_text_align()` starting from
-    // `parent_` (NOT the Label itself — the Label's own value IS
-    // match-parent, which would loop). If no ancestor set a value, fall
-    // back to `left` (the CSS spec default for `text-align`).
-    //
-    // If the parent's resolved value is itself `start`/`end` (encoded as
-    // `left`/`right` here since pulp aliases them at the setter), the
-    // existing LTR-only writing-direction policy applies — same as `auto`.
-    if (effective_text_align == LabelAlign::match_parent) {
-        // Walk ancestor chain manually, skipping any intermediate
-        // View whose own inh_text_align_ is also 5 (match-parent).
-        // `View::inheritable_text_align()` is unsuitable here because
-        // it returns the first SET value, not the first NON-match-parent
-        // value — if the immediate parent is also match-parent, the
-        // walk has to continue. Codex P1 on PR #1879 (2026-05-12).
-        //
-        // Example chain we now resolve correctly:
-        //   grandparent  text-align: center        (slot = 1)
-        //   parent       text-align: match-parent  (slot = 5)
-        //   child Label  text-align: match-parent  → renders center
-        //
-        // Previous code stopped at parent.inheritable_text_align()
-        // returning 5 and degraded to `left`.
-        LabelAlign parent_resolved = LabelAlign::left;
-        for (auto* anc = parent(); anc != nullptr; anc = anc->parent()) {
-            auto inh = anc->inheritable_text_align();
-            if (!inh.has_value()) continue;
-            int v = inh.value();
-            if (v == 5) {
-                // Intermediate match-parent ancestor — keep walking.
-                // Note: `inheritable_text_align()` already walks up
-                // until it finds a SET value. We need the loop here
-                // because the SET value might itself be 5; we want
-                // the first non-5 SET value in the chain.
-                //
-                // To skip past this anc's own slot in the next
-                // iteration, we must continue from its parent and
-                // re-enter inheritable_text_align() — but since
-                // inheritable walks again from there, we need to
-                // explicitly bypass anc's slot. Easiest: peek at
-                // anc->parent() directly and re-loop.
-                continue;
-            }
-            if      (v == 1) parent_resolved = LabelAlign::center;
-            else if (v == 2) parent_resolved = LabelAlign::right;
-            else if (v == 3) parent_resolved = LabelAlign::auto_;
-            else if (v == 4) parent_resolved = LabelAlign::justify;
-            else             parent_resolved = LabelAlign::left;
-            break;
-        }
-        // If every ancestor stored 5 (or no ancestor set anything),
-        // CSS spec says fall back to `start` ≡ `left` under LTR.
-        effective_text_align = parent_resolved;
-    }
-
-    // pulp #1434 — resolve `auto` at paint time. `auto` is CSS
-    // writing-direction-relative: LTR → left, RTL → right. Pulp doesn't
-    // model RTL yet (yoga/direction is still NOT-IMPL), so `auto`
-    // currently degrades to `left`. When RTL lands, this becomes a
-    // lookup against the writing-direction context.
-    if (effective_text_align == LabelAlign::auto_) {
-        effective_text_align = LabelAlign::left;
-    }
-
-    float x = 0;
-    switch (effective_text_align) {
-        case LabelAlign::left:
-        case LabelAlign::auto_:         // unreachable — resolved above; keeps switch exhaustive
-        case LabelAlign::match_parent:  // unreachable — resolved above; keeps switch exhaustive
-            canvas.set_text_align(canvas::TextAlign::left);
-            break;
-        case LabelAlign::center:
-            canvas.set_text_align(canvas::TextAlign::center);
-            x = bounds().width * 0.5f;
-            break;
-        case LabelAlign::right:
-            canvas.set_text_align(canvas::TextAlign::right);
-            x = bounds().width;
-            break;
-        case LabelAlign::justify:
-            // pulp #1434 — emit canvas TextAlign::justify so backends
-            // that wire SkParagraph kJustify can render true justified
-            // text. RecordingCanvas / CG fall back to left-alignment
-            // semantics (no kerning-controlled space distribution).
-            // Anchor x at 0 so the first line's leading edge matches a
-            // left-aligned label.
-            canvas.set_text_align(canvas::TextAlign::justify);
-            break;
-    }
-
-    // pulp #1407 — track the actually-painted single-line string so the
-    // decoration block below measures the truncated text, not the
-    // original. Multi-line keeps using display_text since it paints the
-    // full string across multiple draw calls.
-    std::string draw_text = display_text;
-    if (!multi_line_) {
-        // pulp #1407 — CSS `text-overflow: ellipsis`. Truncate with U+2026
-        // when the measured text exceeds the content-box, regardless of
-        // text-align (CSS truncates at the trailing edge for all three).
-        // UTF-8-safe via codepoint binary-search in truncate_to_width().
-        if (text_overflow_ellipsis())
-            draw_text = truncate_to_width(canvas, display_text, bounds().width);
-        canvas.fill_text(draw_text, x, baseline_y);
-    } else {
-        // pulp #1552 — CSS `line-clamp` / `-webkit-line-clamp`. When the
-        // clamp count is set and the text would emit more lines than
-        // allowed, paint at most N lines and append the U+2026 ellipsis
-        // to the last visible line if any source lines were dropped.
-        // 0 disables clamping (matches CSS spec; spec uses `none`).
-        // visible_lines / source_lines / text_h are computed earlier so
-        // vertical-align positioning reflects the clamped block height.
-        const bool need_ellipsis = (line_clamp_ > 0 && source_lines > line_clamp_);
-
-        // Start from the clamped baseline so the first visible line
-        // sits at the top of the centered/bottom-aligned block.
-        float y = baseline_y;
-        int emitted = 0;
-
-        if (use_shaper_wrap) {
-            // pulp #1737 PR-2 — shaped-layout iteration path. TextShaper
-            // already split display_text into shaped_layout.lines using
-            // the active BreakMode (break-word / anywhere). Iterate
-            // those lines instead of `\n`-splitting. Line-clamp +
-            // ellipsis logic is identical — driven by source_lines
-            // (= shaped_layout.line_count) and visible_lines.
-            for (const auto& shaped_line : shaped_layout.lines) {
-                if (emitted >= visible_lines) break;
-                std::string line = shaped_line.text;
-                if (need_ellipsis && (emitted + 1 == visible_lines)) {
-                    line.append("\xe2\x80\xa6");
-                }
-                canvas.fill_text(line, x, y);
-                y += lh;
-                ++emitted;
-            }
-        } else {
-            // Legacy `\n`-only split path — bit-identical to pre-#1737.
-            // Used when bounds().width is 0 / unbounded (the shaper has
-            // no max_width to break against). Existing consumers see
-            // exactly the previous behavior in that case.
-            size_t pos = 0;
-            while (pos < display_text.size()) {
-                if (emitted >= visible_lines) break;
-                size_t nl = display_text.find('\n', pos);
-                if (nl == std::string::npos) nl = display_text.size();
-                std::string line = display_text.substr(pos, nl - pos);
-                // Last visible line under a clamp that truncated source lines:
-                // append U+2026 (UTF-8: 0xE2 0x80 0xA6) to signal truncation.
-                // pulp #1552 — matches the CSS "block-axis ellipsis" intent
-                // for line-clamp (the spec ties this to text-overflow: ellipsis;
-                // pulp's Label honors that intent without requiring the
-                // separate text-overflow keyword to also be set).
-                if (need_ellipsis && (emitted + 1 == visible_lines)) {
-                    line.append("\xe2\x80\xa6");
-                }
-                canvas.fill_text(line, x, y);
-                y += lh;
-                pos = nl + 1;
-                ++emitted;
-            }
-        }
-    }
-
-    // Text decoration (underline, line-through, overline)
-    if (text_decoration_ != TextDecoration::none) {
-        auto dec_color = has_decoration_color_ ? decoration_color_ : text_color;
-        canvas.set_stroke_color(dec_color);
-        canvas.set_line_width(1.0f);
-        // pulp #1407 — measure the actually-drawn (possibly truncated)
-        // text so a decoration line on an ellipsised label doesn't
-        // escape past the visible glyphs.
-        float text_w = canvas.measure_text(draw_text);
-        float draw_x = x;
-        if (effective_text_align == LabelAlign::center) draw_x = x - text_w * 0.5f;
-        else if (effective_text_align == LabelAlign::right) draw_x = x - text_w;
-
-        if (text_decoration_ == TextDecoration::underline)
-            canvas.stroke_line(draw_x, baseline_y + 2, draw_x + text_w, baseline_y + 2);
-        else if (text_decoration_ == TextDecoration::line_through)
-            canvas.stroke_line(draw_x, baseline_y - effective_font_size * 0.2f, draw_x + text_w, baseline_y - effective_font_size * 0.2f);
-        else if (text_decoration_ == TextDecoration::overline)
-            canvas.stroke_line(draw_x, baseline_y - effective_font_size * 0.7f, draw_x + text_w, baseline_y - effective_font_size * 0.7f);
-    }
-
-    if (vertical) canvas.restore();
-
-    // pulp #1737 (Codex P1 on #1791) — clear font_features at end of
-    // paint so subsequent widgets in the same paint pass don't inherit
-    // this Label's OpenType state. The canvas keeps font_features as
-    // mutable canvas-level state, and TextButton / TextEditor / sibling
-    // Labels that call measure_text/fill_text without setting features
-    // would otherwise pick up tnum/smcp/etc. from the previous
-    // fontVariant-bearing Label, causing unintended typography drift
-    // outside that label's box.
-    canvas.clear_font_features();
-}
+// Label moved to core/view/src/widgets/label.cpp
+// in the 2026-05 Phase 2 (R2-6) batch.
 
 // ── Knob ─────────────────────────────────────────────────────────────────────
 
@@ -1460,765 +827,9 @@ void Icon::paint(canvas::Canvas& canvas) {
     }
 }
 
-// ── ImageView ────────────────────────────────────────────────────────────────
-
-void ImageView::paint(canvas::Canvas& canvas) {
-    auto b = local_bounds();
-
-    if (path_.empty()) {
-        // Placeholder: gray rect with "IMG" text
-        canvas.set_fill_color(resolve_color("bg.surface", canvas::Color::rgba8(50, 50, 60)));
-        canvas.fill_rounded_rect(0, 0, b.width, b.height, 4);
-        canvas.set_fill_color(resolve_color("text.secondary", canvas::Color::rgba8(120, 120, 140)));
-        canvas.set_font("Inter", 10);
-        canvas.set_text_align(canvas::TextAlign::center);
-        canvas.fill_text("IMG", b.width * 0.5f, b.height * 0.5f + 3);
-        return;
-    }
-
-    // Workstream 07 B4: when an ImageCache is attached, consult it.
-    // The cache owns decode + eviction; the view just renders whatever
-    // the cache hands back. A null DecodedImage means decode-in-progress
-    // or permanent failure, in which case we fall through to the
-    // path-as-text placeholder so the UI never shows a blank rect.
-    if (cache_) {
-        const auto* decoded = cache_->get(path_);
-        if (decoded && decoded->native_handle) {
-            // Hand the opaque native handle to the canvas. Backends that
-            // know how to interpret it (CoreGraphics → CGImageRef,
-            // Skia → SkImage*) blit it; others no-op and we fall through
-            // to the filename placeholder. Workstream 07 slice B4/#255.
-            canvas.draw_image(decoded->native_handle, 0, 0, b.width, b.height);
-            return;
-        }
-    }
-
-    // pulp #1658 — decode-on-paint via the canvas's draw_image_from_file
-    // primitive. Skia's implementation uses SkData::MakeFromFileName +
-    // SkImages::DeferredFromEncodedData (deferred decode through SkCodec).
-    // Backends that don't implement this primitive (some headless tests)
-    // return false and we fall through to the filename-as-text placeholder
-    // so authors can still see what URL is set.
-    //
-    // Strip a `file://` prefix if present (set_image_path normalises file
-    // paths to that form so the cache layer above can keep URI-keyed
-    // entries; the canvas primitive expects a bare filesystem path).
-    auto fs_path = path_;
-    static const std::string kFileScheme = "file://";
-    if (fs_path.compare(0, kFileScheme.size(), kFileScheme) == 0) {
-        fs_path = fs_path.substr(kFileScheme.size());
-    }
-
-    // pulp #1737 — honour CSS `object-fit` + `object-position`.
-    // Probe intrinsic image dimensions; if the backend can't measure
-    // (no decode primitive on this platform) fall back to the
-    // pre-#1737 stretch-to-bounds path (= object-fit: fill).
-    float img_w = 0.0f, img_h = 0.0f;
-    bool has_intrinsic =
-        !fs_path.empty() &&
-        canvas.measure_image_from_file(fs_path, img_w, img_h) &&
-        img_w > 0.0f && img_h > 0.0f;
-
-    struct FitRect { float x, y, width, height; };
-    auto compute_fit = [&]() -> std::pair<FitRect, FitRect> {
-        // dst rect (relative to view origin), src rect (in image coords).
-        FitRect dst{0.0f, 0.0f, b.width, b.height};
-        FitRect src{0.0f, 0.0f, img_w, img_h};
-        const std::string& fit = object_fit();
-
-        // Default `fill` (CSS spec): stretch to the box, full src.
-        // Same applies when `object-fit` is unset or unknown — the
-        // pre-#1737 behaviour.
-        if (fit.empty() || fit == "fill") {
-            return {dst, src};
-        }
-
-        // `none` — natural size, no scaling. Crop or letterbox both axes.
-        // Centre by default; refined below by object-position parsing.
-        if (fit == "none") {
-            float dw = std::min(b.width,  img_w);
-            float dh = std::min(b.height, img_h);
-            dst = {(b.width - dw) * 0.5f, (b.height - dh) * 0.5f, dw, dh};
-            float sx = (img_w - dw) * 0.5f;
-            float sy = (img_h - dh) * 0.5f;
-            src = {std::max(sx, 0.0f), std::max(sy, 0.0f), dw, dh};
-            return {dst, src};
-        }
-
-        // `contain` (and `scale-down` when img > box) — letterbox,
-        // preserve aspect ratio, scale so the larger axis hits the box.
-        // `cover` — crop, preserve aspect ratio, scale so the smaller
-        // axis hits the box (the other axis overflows + gets cropped).
-        // `scale-down` chooses between `none` and `contain` based on
-        // which produces the smaller painted region.
-        bool is_contain = (fit == "contain");
-        bool is_cover   = (fit == "cover");
-        bool is_scale_down = (fit == "scale-down");
-
-        if (is_scale_down) {
-            // Per spec: scale-down behaves as `none` if it would shrink
-            // (i.e. the image is smaller than the box on both axes);
-            // otherwise behaves as `contain`.
-            if (img_w <= b.width && img_h <= b.height) {
-                // Recurse via the `none` path semantics.
-                float dw = img_w, dh = img_h;
-                dst = {(b.width - dw) * 0.5f, (b.height - dh) * 0.5f, dw, dh};
-                src = {0, 0, img_w, img_h};
-                return {dst, src};
-            }
-            is_contain = true;
-        }
-
-        if (is_contain || is_cover) {
-            float scale_x = b.width  / img_w;
-            float scale_y = b.height / img_h;
-            float scale = is_cover ? std::max(scale_x, scale_y)
-                                   : std::min(scale_x, scale_y);
-            float dw = img_w * scale;
-            float dh = img_h * scale;
-            dst = {(b.width - dw) * 0.5f, (b.height - dh) * 0.5f, dw, dh};
-            src = {0, 0, img_w, img_h};
-            return {dst, src};
-        }
-
-        // Unknown fit keyword — degrade to fill.
-        return {dst, src};
-    };
-
-    if (has_intrinsic) {
-        auto [dst, src] = compute_fit();
-
-        // pulp #1737 — apply `object-position` as a percentage offset.
-        // CSS spec lets the value be lengths or percentages; the JS
-        // shim normalises to a `<x>% <y>%` two-token string before
-        // routing through setObjectPosition. Anything we can't parse
-        // collapses to "50% 50%" (the spec default), which keeps the
-        // centred-by-default `compute_fit` result.
-        auto parse_object_position = [&](float& px, float& py) {
-            px = 50.0f; py = 50.0f;
-            const std::string& s = object_position();
-            if (s.empty()) return;
-            // Tokenise on whitespace; accept `Npx`, `N%`, or bare numbers.
-            std::vector<std::string> toks;
-            std::string cur;
-            for (char c : s) {
-                if (c == ' ' || c == '\t') {
-                    if (!cur.empty()) { toks.push_back(cur); cur.clear(); }
-                } else {
-                    cur.push_back(c);
-                }
-            }
-            if (!cur.empty()) toks.push_back(cur);
-
-            auto parse_one = [&](const std::string& tok, float box_extent,
-                                  float img_extent, float& out_pct) {
-                // Percentage: linear; CSS object-position percentage
-                // means "<P% of the SLACK> measured from the start".
-                if (!tok.empty() && tok.back() == '%') {
-                    try { out_pct = std::stof(tok.substr(0, tok.size() - 1)); }
-                    catch (...) {}
-                    return;
-                }
-                // Length (px): convert to a percentage of the slack.
-                float n = 0.0f;
-                bool ok = true;
-                try { n = std::stof(tok); } catch (...) { ok = false; }
-                if (!ok) return;
-                float slack = box_extent - img_extent;
-                if (std::abs(slack) < 0.001f) { out_pct = 50.0f; return; }
-                out_pct = (n / slack) * 100.0f;
-            };
-
-            if (toks.size() >= 1) parse_one(toks[0], b.width,  dst.width,  px);
-            if (toks.size() >= 2) parse_one(toks[1], b.height, dst.height, py);
-        };
-
-        if (!object_position().empty()) {
-            float px = 50.0f, py = 50.0f;
-            parse_object_position(px, py);
-            float slack_x = b.width  - dst.width;
-            float slack_y = b.height - dst.height;
-            dst.x = slack_x * (px / 100.0f);
-            dst.y = slack_y * (py / 100.0f);
-        }
-
-        // Route via the source-rect overload when `none` / `cover` / a
-        // non-identity src was computed. The dst-only path stays for
-        // the simple `fill` / `contain` cases (full src rect).
-        bool drawn = false;
-        if (src.x != 0.0f || src.y != 0.0f ||
-            std::abs(src.width  - img_w) > 0.001f ||
-            std::abs(src.height - img_h) > 0.001f) {
-            drawn = canvas.draw_image_from_file_rect(
-                fs_path,
-                src.x, src.y, src.width, src.height,
-                dst.x, dst.y, dst.width, dst.height);
-        } else {
-            drawn = canvas.draw_image_from_file(
-                fs_path, dst.x, dst.y, dst.width, dst.height);
-        }
-        if (drawn) { loaded_ = true; return; }
-    } else if (!fs_path.empty() &&
-               canvas.draw_image_from_file(fs_path, 0, 0, b.width, b.height)) {
-        loaded_ = true;
-        return;
-    }
-
-    // Decode failed (file missing, unsupported codec, or backend doesn't
-    // implement draw_image_from_file). Fall through to the filename
-    // placeholder so authors can debug src wiring without a blank rect.
-    canvas.set_fill_color(resolve_color("bg.surface", canvas::Color::rgba8(50, 50, 60)));
-    canvas.fill_rounded_rect(0, 0, b.width, b.height, 4);
-    canvas.set_fill_color(resolve_color("text.secondary", canvas::Color::rgba8(120, 120, 140)));
-    canvas.set_font("Inter", 9);
-    canvas.set_text_align(canvas::TextAlign::center);
-
-    // Show filename
-    auto name = path_;
-    auto slash = name.rfind('/');
-    if (slash != std::string::npos) name = name.substr(slash + 1);
-    canvas.fill_text(name, b.width * 0.5f, b.height * 0.5f + 3);
-}
-
-// ── Meter ────────────────────────────────────────────────────────────────────
-
-void Meter::set_level(float rms, float peak) {
-    current_rms_ = std::clamp(rms, 0.0f, 1.0f);
-    current_peak_ = std::clamp(peak, 0.0f, 1.0f);
-    ballistics_.display_rms = current_rms_;
-    ballistics_.display_peak = current_peak_;
-}
-
-void Meter::update(float raw_peak, float raw_rms, float dt) {
-    ballistics_.update(raw_peak, raw_rms, dt);
-}
-
-void Meter::paint(canvas::Canvas& canvas) {
-    auto b = local_bounds();
-    bool vert = orientation_ == Orientation::vertical;
-
-    if (render_style_ == WidgetRenderStyle::minimal) {
-        // ── Minimal: gradient bar (green→red) matching design tool appearance ──
-        auto green = resolve_color("accent.success", canvas::Color::rgba8(166, 227, 161));
-        auto red = resolve_color("accent.error", canvas::Color::rgba8(243, 139, 168));
-        // Simple two-stop vertical gradient approximation
-        for (int y = 0; y < static_cast<int>(b.height); ++y) {
-            float t = static_cast<float>(y) / b.height;
-            canvas.set_fill_color(green.interpolate(red, t));
-            canvas.fill_rect(0, static_cast<float>(y), b.width, 1);
-        }
-        // Round corners by clipping (approximate with rounded rect overlay)
-        return;
-    }
-
-    // Background
-    auto bg = resolve_color("control.track", canvas::Color::rgba8(30, 30, 30));
-    canvas.set_fill_color(bg);
-    canvas.fill_rounded_rect(0, 0, b.width, b.height, 2.0f);
-
-    float meter_length = vert ? b.height : b.width;
-
-    // RMS fill (main body)
-    auto rms_color = resolve_color("accent.success", canvas::Color::rgba8(80, 200, 80));
-    float rms_level = ballistics_.display_rms;
-
-    // Color changes at different levels
-    if (rms_level > 0.9f)
-        rms_color = resolve_color("accent.error", canvas::Color::rgba8(240, 60, 60));
-    else if (rms_level > 0.7f)
-        rms_color = resolve_color("accent.warning", canvas::Color::rgba8(240, 180, 60));
-
-    canvas.set_fill_color(rms_color);
-    float fill = rms_level * meter_length;
-
-    if (vert) {
-        canvas.fill_rect(1, b.height - fill, b.width - 2, fill);
-    } else {
-        canvas.fill_rect(0, 1, fill, b.height - 2);
-    }
-
-    // Peak indicator line
-    float peak_level = ballistics_.display_peak;
-    if (peak_level > 0.01f) {
-        auto peak_color = resolve_color("control.thumb", canvas::Color::rgba8(255, 255, 255));
-        canvas.set_stroke_color(peak_color);
-        canvas.set_line_width(1.0f);
-
-        float peak_pos = peak_level * meter_length;
-        if (vert) {
-            float y = b.height - peak_pos;
-            canvas.stroke_line(1, y, b.width - 1, y);
-        } else {
-            canvas.stroke_line(peak_pos, 1, peak_pos, b.height - 1);
-        }
-    }
-
-    // Held peak indicator
-    float held = ballistics_.held_peak;
-    if (held > 0.01f) {
-        auto held_color = canvas::Color::rgba8(255, 100, 100);
-        if (held > 0.9f)
-            held_color = resolve_color("accent.error", canvas::Color::rgba8(255, 50, 50));
-
-        canvas.set_stroke_color(held_color);
-        canvas.set_line_width(2.0f);
-
-        float held_pos = held * meter_length;
-        if (vert) {
-            float y = b.height - held_pos;
-            canvas.stroke_line(0, y, b.width, y);
-        } else {
-            canvas.stroke_line(held_pos, 0, held_pos, b.height);
-        }
-    }
-}
-
-// ── XYPad ────────────────────────────────────────────────────────────────────
-
-void XYPad::update_from_pos(Point pos) {
-    auto b = local_bounds();
-    if (b.width <= 0 || b.height <= 0) return;
-    float new_x = std::clamp(pos.x / b.width, 0.0f, 1.0f);
-    float new_y = std::clamp(1.0f - pos.y / b.height, 0.0f, 1.0f);
-    if (new_x != x_ || new_y != y_) {
-        x_ = new_x;
-        y_ = new_y;
-        if (on_change) on_change(x_, y_);
-    }
-}
-
-void XYPad::on_mouse_down(Point pos) { update_from_pos(pos); }
-void XYPad::on_mouse_drag(Point pos) { update_from_pos(pos); }
-
-void XYPad::paint(canvas::Canvas& canvas) {
-    auto b = local_bounds();
-
-    // Background
-    auto bg = resolve_color("bg.surface", canvas::Color::rgba8(40, 40, 55));
-    canvas.set_fill_color(bg);
-    canvas.fill_rounded_rect(0, 0, b.width, b.height, 4.0f);
-
-    // Grid lines
-    auto grid = resolve_color("control.border", canvas::Color::rgba8(60, 60, 75));
-    canvas.set_stroke_color(grid);
-    canvas.set_line_width(0.5f);
-    canvas.stroke_line(b.width * 0.5f, 0, b.width * 0.5f, b.height);
-    canvas.stroke_line(0, b.height * 0.5f, b.width, b.height * 0.5f);
-
-    // Crosshair position — inset by dot radius so it doesn't clip at edges
-    float dot_r = 5.0f;
-    float cx = dot_r + x_ * (b.width - 2.0f * dot_r);
-    float cy = dot_r + (1.0f - y_) * (b.height - 2.0f * dot_r);
-
-    // Crosshair lines
-    auto hair_color = resolve_color("control.fill", canvas::Color::rgba8(100, 150, 255));
-    canvas.set_stroke_color(hair_color);
-    canvas.set_line_width(1.0f);
-    canvas.stroke_line(cx, 0, cx, b.height);
-    canvas.stroke_line(0, cy, b.width, cy);
-
-    // Thumb dot
-    auto thumb = resolve_color("control.thumb", canvas::Color::rgba8(220, 220, 220));
-    canvas.set_fill_color(thumb);
-    canvas.fill_circle(cx, cy, dot_r);
-
-    // Labels
-    auto text_color = resolve_color("text.secondary", canvas::Color::rgba8(150, 150, 150));
-    canvas.set_fill_color(text_color);
-    canvas.set_font("Inter", 9.0f);
-
-    if (!x_label_.empty()) {
-        canvas.set_text_align(canvas::TextAlign::center);
-        canvas.fill_text(x_label_, b.width * 0.5f, b.height - 6);
-    }
-    if (!y_label_.empty()) {
-        canvas.set_text_align(canvas::TextAlign::left);
-        canvas.fill_text(y_label_, 2, 10);
-    }
-}
-
-// ── WaveformView ─────────────────────────────────────────────────────────────
-
-size_t WaveformView::find_trigger_index(const float* samples, size_t count,
-                                         TriggerMode mode) {
-    if (mode == TriggerMode::free_run || count < 2) return 0;
-    const bool want_rising = (mode == TriggerMode::rising_zero);
-    for (size_t i = 1; i < count; ++i) {
-        float prev = samples[i - 1];
-        float curr = samples[i];
-        if (want_rising) {
-            if (prev <= 0.0f && curr > 0.0f) return i;
-        } else {
-            if (prev >= 0.0f && curr < 0.0f) return i;
-        }
-    }
-    return 0;
-}
-
-void WaveformView::apply_trigger() {
-    if (trigger_mode_ == TriggerMode::free_run || samples_.size() < 2) return;
-    size_t idx = find_trigger_index(samples_.data(), samples_.size(), trigger_mode_);
-    if (idx == 0) return;  // no crossing — leave as-is
-    std::rotate(samples_.begin(),
-                samples_.begin() + static_cast<std::ptrdiff_t>(idx),
-                samples_.end());
-}
-
-void WaveformView::set_data(const float* samples, size_t count) {
-    samples_.assign(samples, samples + count);
-    apply_trigger();
-}
-
-void WaveformView::set_data(std::vector<float> samples) {
-    samples_ = std::move(samples);
-    apply_trigger();
-}
-
-void WaveformView::paint(canvas::Canvas& canvas) {
-    auto b = local_bounds();
-
-    // Background
-    auto bg = resolve_color("bg.surface", canvas::Color::rgba8(30, 30, 40));
-    canvas.set_fill_color(bg);
-    canvas.fill_rounded_rect(0, 0, b.width, b.height, 2.0f);
-
-    if (samples_.empty()) return;
-
-    // Center line
-    auto center_color = resolve_color("waveform.grid", canvas::Color::rgba8(50, 50, 60));
-    canvas.set_stroke_color(center_color);
-    canvas.set_line_width(0.5f);
-    float cy = b.height * 0.5f;
-    canvas.stroke_line(0, cy, b.width, cy);
-
-    auto wave_color = resolve_color("waveform.line", canvas::Color::rgba8(100, 180, 250));
-    auto fill_color = resolve_color("waveform.fill", canvas::Color::rgba(wave_color.r, wave_color.g, wave_color.b, 56.0f/255.0f));
-
-    // GPU-accelerated waveform rendering via SkSL shader
-    canvas::Canvas::WaveformStyle style;
-    style.line_color = wave_color;
-    style.fill_color = fill_color;
-    style.line_thickness = 2.0f;
-    style.show_fill = true;
-    style.fill_center = 0.5f;
-
-    canvas.draw_waveform(samples_.data(), samples_.size(), 0, 0, b.width, b.height, style);
-}
-
-// ── SpectrumView ─────────────────────────────────────────────────────────────
-
-void SpectrumView::set_spectrum(const float* magnitudes_db, size_t bin_count) {
-    bins_.assign(magnitudes_db, magnitudes_db + bin_count);
-}
-
-void SpectrumView::set_spectrum(std::vector<float> magnitudes_db) {
-    bins_ = std::move(magnitudes_db);
-}
-
-void SpectrumView::paint(canvas::Canvas& canvas) {
-    auto b = local_bounds();
-
-    // Background
-    auto bg = resolve_color("bg.surface", canvas::Color::rgba8(25, 25, 35));
-    canvas.set_fill_color(bg);
-    canvas.fill_rounded_rect(0, 0, b.width, b.height, 2.0f);
-
-    if (bins_.empty()) return;
-
-    float db_range = max_db_ - min_db_;
-    if (db_range <= 0) return;
-
-    auto spectrum_color = resolve_color("waveform.line", canvas::Color::rgba8(100, 180, 250));
-
-    if (style_ == Style::bars) {
-        canvas.set_fill_color(spectrum_color);
-        float bar_width = b.width / static_cast<float>(bins_.size());
-
-        for (size_t i = 0; i < bins_.size(); ++i) {
-            float norm = std::clamp((bins_[i] - min_db_) / db_range, 0.0f, 1.0f);
-            float bar_h = norm * b.height;
-            float x = i * bar_width;
-            canvas.fill_rect(x, b.height - bar_h, std::max(bar_width - 1, 1.0f), bar_h);
-        }
-    } else {
-        // Line or filled style — use GPU waveform shader for anti-aliased rendering
-        auto fill = resolve_color("waveform.fill", canvas::Color::rgba(spectrum_color.r, spectrum_color.g, spectrum_color.b, 60.0f/255.0f));
-
-        // Normalize dB values to -1..1 for the waveform shader (0 dB → top, min_db → bottom)
-        std::vector<float> normalized(bins_.size());
-        for (size_t i = 0; i < bins_.size(); ++i) {
-            float norm = std::clamp((bins_[i] - min_db_) / db_range, 0.0f, 1.0f);
-            normalized[i] = norm * 2.0f - 1.0f;  // Map 0..1 → -1..1
-        }
-
-        canvas::Canvas::WaveformStyle ws;
-        ws.line_color = spectrum_color;
-        ws.fill_color = fill;
-        ws.line_thickness = 1.5f;
-        ws.show_fill = (style_ == Style::filled);
-        ws.fill_center = 1.0f;  // Fill from bottom
-
-        canvas.draw_waveform(normalized.data(), normalized.size(), 0, 0, b.width, b.height, ws);
-    }
-
-    // Frequency grid lines (approximate positions)
-    auto grid = resolve_color("waveform.grid", canvas::Color::rgba8(50, 50, 65));
-    canvas.set_stroke_color(grid);
-    canvas.set_line_width(0.5f);
-
-    // Horizontal dB grid lines at -20, -40, -60
-    for (float db : {-20.0f, -40.0f, -60.0f}) {
-        float norm = std::clamp((db - min_db_) / db_range, 0.0f, 1.0f);
-        float y = b.height - norm * b.height;
-        canvas.stroke_line(0, y, b.width, y);
-    }
-}
-
-// ── Panel ────────────────────────────────────────────────────────────────────
-
-void Panel::paint(canvas::Canvas& canvas) {
-    auto b = local_bounds();
-    auto bg = resolve_color(bg_token_, canvas::Color::rgba8(45, 45, 60));
-    canvas.set_fill_color(bg);
-    // pulp #1663 — resolve the effective uniform radius (px or %).
-    const float eff_radius = effective_corner_radius(b.width, b.height);
-    canvas.fill_rounded_rect(0, 0, b.width, b.height, eff_radius);
-
-    if (border_width_ > 0) {
-        auto border = resolve_color(border_token_, canvas::Color::rgba8(80, 80, 100));
-        canvas.set_stroke_color(border);
-        canvas.set_line_width(border_width_);
-        // Inset stroke by half its width so the center lands on pixel boundaries
-        // (Visage-style half-pixel alignment for crisper edges)
-        float inset = border_width_ * 0.5f;
-        canvas.stroke_rounded_rect(inset, inset,
-                                    b.width - border_width_, b.height - border_width_,
-                                    std::max(0.0f, eff_radius - inset));
-    }
-}
-
-// ── SpectrogramView ──────────────────────────────────────────────────────────
-
-void SpectrogramView::configure(int history_columns, int freq_rows,
-                                 signal::ColorRamp ramp, float min_db, float max_db) {
-    buffer_.configure(history_columns, freq_rows);
-    mapper_.set_ramp(ramp);
-    min_db_ = min_db;
-    max_db_ = max_db;
-    configured_ = true;
-}
-
-void SpectrogramView::push_spectrum(const float* magnitudes_db, int num_bins) {
-    if (!configured_) {
-        // Auto-configure with sensible defaults
-        configure(256, num_bins);
-    }
-    buffer_.push_column(magnitudes_db, num_bins, mapper_, min_db_, max_db_);
-}
-
-void SpectrogramView::paint(canvas::Canvas& canvas) {
-    auto b = local_bounds();
-
-    // Background
-    auto bg = resolve_color("bg.surface", canvas::Color::rgba8(10, 10, 15));
-    canvas.set_fill_color(bg);
-    canvas.fill_rect(0, 0, b.width, b.height);
-
-    if (!configured_ || buffer_.frames_written() == 0) return;
-
-    int buf_w = buffer_.width();
-    int buf_h = buffer_.height();
-    float px_w = b.width / buf_w;
-    float px_h = b.height / buf_h;
-
-    int start_col = buffer_.write_column(); // oldest column
-
-    for (int col = 0; col < buf_w; ++col) {
-        int src_col = (start_col + col) % buf_w;
-        float x = col * px_w;
-
-        for (int row = 0; row < buf_h; ++row) {
-            auto c = buffer_.pixels()[row * buf_w + src_col];
-            // Flip vertically: low freq at bottom
-            float y = b.height - (row + 1) * px_h;
-            canvas.set_fill_color(canvas::Color::rgba(c.r, c.g, c.b, c.a));
-            canvas.fill_rect(x, y, px_w + 0.5f, px_h + 0.5f);
-        }
-    }
-}
-
-// ── MultiMeter ──────────────────────────────────────────────────────────────
-
-void MultiMeter::set_channel_count(int count) {
-    ballistics_.num_channels = std::min(count, signal::kMaxMeterChannels);
-}
-
-void MultiMeter::update(const signal::MultiChannelMeterData& data, float dt) {
-    ballistics_.update(data, dt);
-}
-
-void MultiMeter::paint(canvas::Canvas& canvas) {
-    auto b = local_bounds();
-    int num_ch = ballistics_.num_channels;
-    if (num_ch <= 0) return;
-
-    // Background
-    auto bg = resolve_color("control.track", canvas::Color::rgba8(30, 30, 30));
-    canvas.set_fill_color(bg);
-    canvas.fill_rounded_rect(0, 0, b.width, b.height, 2.0f);
-
-    bool vert = (layout_ == Layout::vertical);
-    float gap = 2.0f;
-
-    // Calculate per-channel dimensions
-    float ch_size;
-    if (vert)
-        ch_size = (b.width - gap * (num_ch - 1)) / num_ch;
-    else
-        ch_size = (b.height - gap * (num_ch - 1)) / num_ch;
-
-    for (int ch = 0; ch < num_ch; ++ch) {
-        auto& bc = ballistics_.channels[ch];
-        float meter_length = vert ? b.height : b.width;
-
-        // Channel position
-        float x0, y0, cw, ch_h;
-        if (vert) {
-            x0 = ch * (ch_size + gap);
-            y0 = 0;
-            cw = ch_size;
-            ch_h = b.height;
-        } else {
-            x0 = 0;
-            y0 = ch * (ch_size + gap);
-            cw = b.width;
-            ch_h = ch_size;
-        }
-
-        // RMS fill
-        auto rms_color = resolve_color("accent.success", canvas::Color::rgba8(80, 200, 80));
-        float rms_level = bc.display_rms;
-        if (rms_level > 0.9f)
-            rms_color = resolve_color("accent.error", canvas::Color::rgba8(240, 60, 60));
-        else if (rms_level > 0.7f)
-            rms_color = resolve_color("accent.warning", canvas::Color::rgba8(240, 180, 60));
-
-        canvas.set_fill_color(rms_color);
-        float fill = rms_level * meter_length;
-        if (vert)
-            canvas.fill_rect(x0, ch_h - fill, cw, fill);
-        else
-            canvas.fill_rect(x0, y0, fill, ch_h);
-
-        // Peak indicator
-        float peak_level = bc.display_peak;
-        if (peak_level > 0.01f) {
-            canvas.set_stroke_color(canvas::Color::rgba8(255, 255, 255));
-            canvas.set_line_width(1.0f);
-            float peak_pos = peak_level * meter_length;
-            if (vert) {
-                float y = ch_h - peak_pos;
-                canvas.stroke_line(x0, y, x0 + cw, y);
-            } else {
-                canvas.stroke_line(peak_pos, y0, peak_pos, y0 + ch_h);
-            }
-        }
-
-        // Held peak
-        float held = bc.held_peak;
-        if (held > 0.01f) {
-            auto held_color = held > 0.9f
-                ? resolve_color("accent.error", canvas::Color::rgba8(255, 50, 50))
-                : canvas::Color::rgba8(255, 100, 100);
-            canvas.set_stroke_color(held_color);
-            canvas.set_line_width(2.0f);
-            float held_pos = held * meter_length;
-            if (vert) {
-                float y = ch_h - held_pos;
-                canvas.stroke_line(x0, y, x0 + cw, y);
-            } else {
-                canvas.stroke_line(held_pos, y0, held_pos, y0 + ch_h);
-            }
-        }
-
-        // Clip indicator
-        if (bc.clip_indicator) {
-            canvas.set_fill_color(resolve_color("accent.error", canvas::Color::rgba8(255, 0, 0)));
-            if (vert)
-                canvas.fill_rect(x0, 0, cw, 3.0f);
-            else
-                canvas.fill_rect(cw - 3.0f, y0, 3.0f, ch_h);
-        }
-    }
-}
-
-// ── CorrelationMeter ────────────────────────────────────────────────────────
-
-void CorrelationMeter::update(float correlation, float dt) {
-    float target = std::clamp(correlation, -1.0f, 1.0f);
-    float coeff = 1.0f - std::exp(-dt / 0.05f); // 50ms smoothing
-    display_correlation_ += (target - display_correlation_) * coeff;
-}
-
-void CorrelationMeter::paint(canvas::Canvas& canvas) {
-    auto b = local_bounds();
-
-    // Background
-    auto bg = resolve_color("control.track", canvas::Color::rgba8(30, 30, 30));
-    canvas.set_fill_color(bg);
-    canvas.fill_rounded_rect(0, 0, b.width, b.height, 2.0f);
-
-    // Center line (0.0 correlation)
-    float center_x = b.width * 0.5f;
-    canvas.set_stroke_color(canvas::Color::rgba8(80, 80, 80));
-    canvas.set_line_width(1.0f);
-    canvas.stroke_line(center_x, 0, center_x, b.height);
-
-    // -1 and +1 labels position markers
-    float quarter = b.width * 0.25f;
-    canvas.set_stroke_color(canvas::Color::rgba8(50, 50, 50));
-    canvas.stroke_line(quarter, 0, quarter, b.height);
-    canvas.stroke_line(b.width - quarter, 0, b.width - quarter, b.height);
-
-    // Correlation indicator
-    // Map -1..+1 to 0..width, inset by half bar width to prevent edge clipping
-    float norm = (display_correlation_ + 1.0f) * 0.5f; // 0..1
-    float bar_width = std::max(4.0f, b.width * 0.02f);
-    float usable = b.width - bar_width;
-    float indicator_x = bar_width * 0.5f + norm * usable;
-
-    // Color: green at +1, yellow at 0, red at -1
-    canvas::Color indicator_color;
-    if (display_correlation_ > 0.0f) {
-        // Green to yellow
-        float t = display_correlation_;
-        indicator_color = canvas::Color::rgba8(
-            static_cast<uint8_t>(240 * (1.0f - t)),
-            static_cast<uint8_t>(200 + 55 * t),
-            static_cast<uint8_t>(60 * (1.0f - t)));
-    } else {
-        // Yellow to red
-        float t = -display_correlation_;
-        indicator_color = canvas::Color::rgba8(
-            static_cast<uint8_t>(240),
-            static_cast<uint8_t>(200 * (1.0f - t)),
-            static_cast<uint8_t>(60 * (1.0f - t)));
-    }
-
-    // Draw indicator bar
-    canvas.set_fill_color(indicator_color);
-    canvas.fill_rounded_rect(indicator_x - bar_width * 0.5f, 1,
-                              bar_width, b.height - 2, 2.0f);
-
-    // Fill from center to indicator
-    auto fill_color = indicator_color;
-    fill_color.a = 80;
-    canvas.set_fill_color(fill_color);
-    if (indicator_x > center_x) {
-        canvas.fill_rect(center_x, 2, indicator_x - center_x, b.height - 4);
-    } else {
-        canvas.fill_rect(indicator_x, 2, center_x - indicator_x, b.height - 4);
-    }
-}
+// Visualizers (ImageView / Meter / XYPad / WaveformView /
+// SpectrumView / Panel / SpectrogramView / MultiMeter /
+// CorrelationMeter) moved to core/view/src/widgets/visualizers.cpp
+// in the 2026-05 Phase 2 (R2-6) batch.
 
 } // namespace pulp::view
